chore: converted aurora-bg into rsc, renamed spotlight
