+++ src/components/ui/aurora-background.tsx
@@ -1,5 +1,3 @@
-'use client'
-
 import React, { ReactNode } from 'react'
 
 import { cn } from '@/lib/utils'
